requirements.txt
requests==2.25.1
regex==2021.9.30
matplotlib==3.5.1
pandas==1.4.3
Pillow==9.2.0
wordcloud==1.8.1